Move files and rename classes.
